burn_jit

Module kernel

Kernel module

Modules

conv

Convolution kernels

interpolate

Interpolation kernels

matmul

Matmul kernels

pool

Pooling kernels

prng

Pseudo-random number generator kernels

quantization

Quantization operations

reduce

Reduction algorithms

Enums

MaskFillStrategy

Define how to run the mask fill kernel.

MaskWhereStrategy

Define how to run the mask where kernel.

Constants

PLANE_DIM_APPROX

An approximation of the plane dimension.

Traits

Kernel

Implement this trait to create a kernel definition.

Functions

bool_cast

Cast a bool tensor to the given element type.

cast

Cast a tensor to the given element type.

into_contiguous

Make a jit tensor contiguous.

mask_fill

Execute the mask fill kernel with the given strategy.

mask_where

Execute the mask where kernel with the given strategy.