burn_jit::kernel

Module quantization

Quantization operations

Structs

QParams

Quantization parameters.

QParamsCompilationArg

QParamsExpand

QParamsLaunch

Functions

dequantize

Convert the tensor back to a higher precision data type.

quantize

Convert the tensor to a lower precision data type based on the quantization scheme and parameters.

Type Aliases

QTensor

Quantized tensor representation.