burn_jit::kernel

Module reduce

Reduction algorithms

Structs

ArgMax

Compute the coordinate of the maximum item returning the smallest coordinate in case of equality.

ArgMin

Compute the coordinate of the maximum item returning the smallest coordinate in case of equality.

Mean

Prod

Reduce

ReduceAutotuneKey

Autotune key representative of redue versions

ReduceOps

ReducePlane

ReduceShared

ReduceSharedPlane

Sum

Enums

ReduceStrategy

Select a strategy to perform a reduction.

Functions

autotune_reduce

Executes autotune on reduce operations.

init_reduce_output

Creates an empty output tensor with the proper shape and decreasing strides to reduce the given axis of input or return None if axis is out-of-bound.

reduce

Reduce all elements of the input tensor using the instruction Rd and the given Strategy.

reduce_dim

Reduce the given axis of the input tensor using the instruction Rd and the given Strategy.