burn_jit::kernel::reduce

Struct ArgMax

pub struct ArgMax;

Compute the coordinate of the maximum item returning the smallest coordinate in case of equality.

Implementations

impl ArgMax

pub fn choose_argmax<N>( items0: Line<N>, coordinates0: Line<u32>, items1: Line<N>, coordinates1: Line<u32>, ) -> (Line<N>, Line<u32>)
where N: Numeric,

Compare two pairs of items and coordinates and return a new pair where each element in the lines is the maximal item with its coordinate. In case of equality, the lowest coordinate is selected.

pub fn __expand_choose_argmax<N>( context: &mut CubeContext, items0: <Line<N> as CubeType>::ExpandType, coordinates0: <Line<u32> as CubeType>::ExpandType, items1: <Line<N> as CubeType>::ExpandType, coordinates1: <Line<u32> as CubeType>::ExpandType, ) -> <(Line<N>, Line<u32>) as CubeType>::ExpandType
where N: Numeric,