burn_jit::tensor

Struct JitTensor

pub struct JitTensor<R: JitRuntime> {
    pub client: ComputeClient<R::Server, R::Channel>,
    pub handle: Handle,
    pub shape: Shape,
    pub device: R::Device,
    pub strides: Vec<usize>,
    /* private fields */
}

The basic tensor primitive struct.

Fields
client: ComputeClient<R::Server, R::Channel>

Compute client for the runtime.

handle: Handle

The buffer where the data are stored.

shape: Shape

The shape of the tensor.

device: R::Device

The device of the tensor.

strides: Vec<usize>

The strides of the tensor.

Implementations

impl<R: JitRuntime> JitTensor<R>

pub fn new( client: ComputeClient<R::Server, R::Channel>, handle: Handle, shape: Shape, device: R::Device, strides: Vec<usize>, dtype: DType, ) -> Self

Constructs a new JitTensor.

impl<R> JitTensor<R>
where R: JitRuntime,

pub fn new_contiguous( client: ComputeClient<R::Server, R::Channel>, device: R::Device, shape: Shape, handle: Handle, dtype: DType, ) -> Self

Create a new tensor with a contiguous memory layout.

pub fn to_client( &self, client: ComputeClient<R::Server, R::Channel>, device: R::Device, ) -> Self

Change the context of the current tensor and return the newly transferred tensor.

pub fn as_handle_ref(&self) -> TensorHandleRef<'_, R>

Return the reference to a tensor handle.

pub fn as_tensor_arg<'a, E: JitElement>( &'a self, vectorisation: u8, ) -> TensorArg<'a, R>

Return the reference to a tensor argument.

pub fn as_array_arg<E: JitElement>(&self, vectorisation: u8) -> ArrayArg<'_, R>

Return the reference to an array argument.

pub fn copy(&self) -> Self

Copy the current tensor.

pub fn can_mut(&self) -> bool

Check if the tensor is safe to mutate.

pub fn assert_is_on_same_device(&self, other: &Self)

Assert that both tensors are on the same device.

pub fn is_contiguous(&self) -> bool

Check if the current tensor is contiguous.

pub fn is_contiguous_buffer(&self) -> bool

Check if the current tensor has a contiguous backing buffer (no overlap and no empty memory regions within the shape).